add header and fix tracker events

+++ Sc2ReplayAnalyzer/Decoder/EventID.cs
@@ -5,7 +5,7 @@ namespace Sc2ReplayAnalyzer.Decoder;
 
 internal record class TrackerEventPair
 {
-    public uint Delta;
+    public long Delta;
     public ReplayTrackerEEventId TrackerEventID;
 }
